refactor(ui): TMS-2587 add user container
Adding a user management container so that the tenant management container doesn't need to be so complicated.
