+++ frontend/src/components/auth/LoginContainer.vue
@@ -3,7 +3,11 @@ import { watchEffect } from 'vue'
 
 import { useAuthStore } from '@/stores'
 
+// --- Store and Composable Setup ----------------------------------------------
+
 const authStore = useAuthStore()
+
+// --- Watchers and Effects ----------------------------------------------------
 
 watchEffect(() => {
   if (!authStore.isAuthenticated) {
